requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy
psycopg2-binary
alembic
pydantic
pydantic[email]
pydantic-settings
bcrypt
python-jose
python-dotenv
passlib[bcrypt]